turns out the `plumber` was sleeping on the job so fixed that and now works again

+++ bin/tools/documentation.sh
@@ -8,11 +8,14 @@ __buildDocumentationBuildDirectory() {
   local home="$1" && shift
   local aa=() target
 
+  # Fill in any missing files
   local prefix="$home/documentation/source"
   while read -r markdownFile; do
     target="$home/documentation/.docs${markdownFile#"$prefix"}"
     __catchEnvironment "$handler" muzzle fileDirectoryRequire "$target" || return $?
-    __catchEnvironment "$handler" cp "$markdownFile" "$target" || return $?
+    if [ ! -f "$target" ]; then
+      __catchEnvironment "$handler" cp "$markdownFile" "$target" || return $?
+    fi
   done < <(find "$prefix" -name '*.md' ! -path '*/tools/*')
 
   source="$home/documentation/source/tools"
@@ -80,14 +83,44 @@ __mkdocsConfiguration() {
     __catch "$handler" buildEnvironmentLoad "$token" || return $?
     export "${token?}"
   done < <(mapTokens <"$source")
-  __catch "$handler" mapEnvironment <"$source" >"$target" || return $?
+  version="$version" __catch "$handler" mapEnvironment <"$source" >"$target" || return $?
+}
+
+# Argument: pass - Boolean. true if version should be a variable, false if it should be set to a value
+__checkVersionVariable() {
+  local handler="$1" pass="$2" line="$3"
+
+  local home
+  home=$(__catch "$handler" buildHome) || return $?
+  checkFile="$home/documentation/.docs/index.md"
+  if [ -f "$checkFile" ]; then
+    local title
+    title=$(basename "$checkFile")
+    if grep -q '{version}' <"$checkFile"; then
+      [ "$pass" = false ] || __throwEnvironment "$handler" "Version is not a variable ($pass) and should be (line $line)" || returnUndo $? dumpPipe --lines 3 "$title" <"$checkFile" || return $?
+    else
+      [ "$pass" = true ] || __throwEnvironment "$handler" "Version is a {variable} ($pass) and should NOT be (line $line)" || returnUndo $? dumpPipe --lines 3 "$title" <"$checkFile" || return $?
+    fi
+  else
+    [ "$pass" = false ] || __throwEnvironment "$handler" "$(decorate file "$checkFile") is missing and should exist (line $line)" || return $?
+  fi
 }
 
 # Build the build documentation
 # fn: {base}
+# Argument: --none - Flag. Turn off all updates.
 # Argument: --templates-only - Flag. Just template identical updates.
+# Argument: --templates - Flag. Enable template updates.
+# Argument: --no-templates - Flag. Disable template updates.
+# Argument: --derived - Flag. Enable derived files updates.
+# Argument: --no-derived - Flag. Disable derived files updates.
 # Argument: --derived-only - Flag. Just derived files only.
-# Argument: --reference-only - Flag. Just tools documentation.
+# Argument: --reference - Flag. Enable reference file updates.
+# Argument: --no-reference - Flag. Disable reference file updates.
+# Argument: --reference-only - Flag. Reference file updates.
+# Argument: --mkdocs - Flag. Enable documentation generation.
+# Argument: --no-mkdocs - Flag. Disable documentation generation.
+# Argument: --mkdocs-only - Flag. Documentation generation only.
 # Argument: --clean - Flag. Clean caches.
 # Argument: --verbose - Flag. Clean caches.
 # Argument: --filter filters ... - DashDashDelimitedArguments. Arguments to filter which reference files are updated.
@@ -102,7 +135,8 @@ buildDocumentationBuild() {
   export APPLICATION_NAME
 
   local da=() ea=()
-  local cleanFlag=false updateDerived=true updateTemplates=true updateReference=true makeDocumentation=true
+  local cleanFlag=false updateDerived=true updateTemplates=false updateReference=true makeDocumentation=true
+  local verboseFlag=false
 
   # _IDENTICAL_ argumentNonBlankLoopHandler 6
   local __saved=("$@") __count=$#
@@ -113,45 +147,24 @@ buildDocumentationBuild() {
     case "$argument" in
     # _IDENTICAL_ helpHandler 1
     --help) "$handler" 0 && return $? || return $? ;;
-    --templates-only)
-      updateDerived=false
-      updateTemplates="true"
-      updateReference=false
-      makeDocumentation=false
-      ;;
-    --derived-only)
-      updateDerived="true"
-      updateTemplates=false
-      updateReference=false
-      makeDocumentation=false
-      ;;
-    --reference-only)
-      updateDerived=false
-      updateTemplates=false
-      updateReference="true"
-      makeDocumentation=false
-      ;;
-    --mkdocs-only)
-      updateDerived=false
-      updateTemplates=false
-      updateReference=false
-      makeDocumentation="true"
-      ;;
-    --clean)
-      cleanFlag=true
-      ;;
-    --verbose)
-      da+=("$argument")
-      ea+=("$argument")
-      ;;
-    --filter)
-      da+=("$argument")
-      while [ $# -gt 0 ] && [ "$1" != "--" ]; do da+=("$1") && shift; done
-      ;;
-    --force)
-      da+=("$argument")
-      ea+=("$argument")
-      ;;
+    --none) updateDerived=false && updateTemplates=false && updateReference=false && makeDocumentation=false ;;
+    --all) updateDerived=true && updateTemplates=true && updateReference=true && makeDocumentation=true ;;
+    --templates) updateTemplates=true ;;
+    --no-templates) updateTemplates=false ;;
+    --templates-only) updateDerived=false && updateTemplates="true" && updateReference=false && makeDocumentation=false ;;
+    --derived) updateDerived=true ;;
+    --no-derived) updateDerived=false ;;
+    --derived-only) updateDerived="true" && updateTemplates=false && updateReference=false && makeDocumentation=false ;;
+    --reference) updateReference=true ;;
+    --no-reference) updateReference=false ;;
+    --reference-only) updateDerived=false && updateTemplates=false && updateReference="true" && makeDocumentation=false ;;
+    --mkdocs) makeDocumentation=true ;;
+    --no-mkdocs) makeDocumentation=false ;;
+    --mkdocs-only) updateDerived=false && updateTemplates=false && updateReference=false && makeDocumentation="true" ;;
+    --clean) cleanFlag=true ;;
+    --verbose) da+=("$argument") && ea+=("$argument") && verboseFlag=true ;;
+    --filter) da+=("$argument") && while [ $# -gt 0 ] && [ "$1" != "--" ]; do da+=("$1") && shift; done ;;
+    --force) da+=("$argument") && ea+=("$argument") ;;
     *)
       # _IDENTICAL_ argumentUnknownHandler 1
       __throwArgument "$handler" "unknown #$__index/$__count \"$argument\" ($(decorate each code -- "${__saved[@]}"))" || return $?
@@ -160,6 +173,17 @@ buildDocumentationBuild() {
     shift
   done
 
+  if $verboseFlag; then
+    decorate pair "Update Templates" "$updateTemplates"
+    decorate pair "Update Derived" "$updateDerived"
+    decorate pair "Update Reference" "$updateReference"
+    decorate pair "Make Documentation" "$makeDocumentation"
+  fi
+  local version
+  version=$(hookVersionCurrent)
+  local timestamp
+  timestamp="$(date -u "+%F %T") UTC"
+
   # Greeting
   __catch "$handler" buildEnvironmentLoad APPLICATION_NAME || return $?
   statusMessage lineFill . "$(decorate info "${APPLICATION_NAME} documentation started on $(decorate value "$(date +"%F %T")")") "
@@ -167,15 +191,25 @@ buildDocumentationBuild() {
 
   # --clean
   if $cleanFlag; then
+    ! $verboseFlag || statusMessage decorate info "Cleaning documentation ..."
     # Clean env cache
     __catch "$handler" documentationBuildEnvironment --clean || return $?
     # Clean reference cache
     __buildDocumentationBuildDirectory "$handler" "$home" --clean || return $?
+    __checkVersionVariable "$handler" false "$LINENO" || return $?
     return 0
   fi
 
+  local targetHome="$home/documentation/.docs"
+
   # Ensure we have our target
-  __catchEnvironment "$handler" muzzle directoryRequire "$home/documentation/.docs" || return $?
+  __catchEnvironment "$handler" muzzle directoryRequire "$targetHome" || return $?
+
+  if ! __checkVersionVariable "$handler" false "$LINENO" 2>/dev/null; then
+    ! $verboseFlag || statusMessage decorate warning "Removing index.md which has incorrect version, derived is automatic"
+    __catchEnvironment "$handler" rm -f "$targetHome/index.md" || return $?
+    updateDerived=true
+  fi
 
   # Templates should be up-to-date if making documentation
   if ! $updateTemplates && $makeDocumentation; then
@@ -184,30 +218,27 @@ buildDocumentationBuild() {
     newestDocs=$(directoryNewestFile "$home/documentation/source")
     if fileIsNewest "$newestTemplate" "$newestDocs"; then
       updateTemplates=true
+      ! $verboseFlag || statusMessage decorate info "Templates were changed, update templates is now automatic"
     fi
   fi
 
+  __checkVersionVariable "$handler" false "$LINENO" || return $?
+
   if $updateTemplates; then
-    statusMessage decorate notice "Updating document templates ..."
+    statusMessage --last decorate notice "Updating document templates ..."
     documentationTemplateUpdate "$home/documentation/source" "$home/documentation/template" || return $?
   fi
 
   if $updateDerived; then
     local file
 
-    statusMessage decorate notice "Updating release page ..."
+    statusMessage --last decorate notice "Updating release page ..."
     __buildDocumentationBuildRelease "$handler" "$home" || return $?
 
-    statusMessage decorate notice "Updating mkdocs.yml ..."
+    __checkVersionVariable "$handler" false "$LINENO" || return $?
 
-    __catchEnvironment "$handler" muzzle pushd "./documentation" || return $?
-    __mkdocsConfiguration "$handler" || return $?
-    __catchEnvironment "$handler" muzzle popd || return $?
+    local sourceHome="$home/documentation/source"
 
-    local sourceHome="$home/documentation/source" targetHome="$home/documentation/.docs"
-    local version timestamp
-
-    version=$(hookVersionCurrent) timestamp="$(date -u "+%F %T") UTC"
     while IFS="" read -r file; do
       file=${file#"$sourceHome"}
       statusMessage decorate notice "Copying $file ..."
@@ -225,6 +256,7 @@ buildDocumentationBuild() {
     done < <(
       find "$sourceHome" -type f -name "*.md" ! -path "*/tools/*" ! -path "*/env/*" -print0 | xargs -0 grep -l '{[A-Za-z][^]!\[}]*}'
     )
+    __checkVersionVariable "$handler" true "$LINENO" || return $?
 
     # Coding
     local example
@@ -232,16 +264,22 @@ buildDocumentationBuild() {
     example="$(decorate wrap "    " <"$home/bin/build/tools/example.sh")" || __throwEnvironment "$handler" "generating example" || return $?
     example="$example" __catch "$handler" mapEnvironment <"$home/documentation/source/guide/coding.md" >"$home/documentation/.docs/guide/coding.md" || return $?
 
-    statusMessage decorate notice "Updating env/index.md ..."
+    statusMessage --last decorate notice "Updating env/index.md ..."
     __catch "$handler" documentationBuildEnvironment --verbose "${ea[@]+"${ea[@]}"}" || return $?
+
+    __checkVersionVariable "$handler" true "$LINENO" || return $?
   fi
 
   if "$updateReference"; then
+    statusMessage --last decorate notice "Updating reference ..."
     __buildDocumentationBuildDirectory "$handler" "$home" "$@" "${da[@]+"${da[@]}"}" || return $?
+
+    __checkVersionVariable "$handler" true "$LINENO"
   fi
 
   if "$makeDocumentation"; then
     if ! whichExists mkdocs; then
+      statusMessage --last decorate notice "Installing python and mkdocs ..."
       __catchEnvironment "$handler" pythonInstall || return $?
 
       if [ ! -d "$home/.venv" ]; then
@@ -271,9 +309,14 @@ buildDocumentationBuild() {
       __catchEnvironment "$handler" source "$home/.venv/bin/activate" || return $?
     fi
     __catchEnvironment "$handler" muzzle pushd "./documentation" || return $?
-    __mkdocsConfiguration "$handler" || return $?
+    statusMessage --last decorate notice "Updating mkdocs.yml ..."
 
+    __checkVersionVariable "$handler" true "$LINENO" || return $?
+
+    timestamp="$timestamp" version="$version" __mkdocsConfiguration "$handler" || return $?
+    __checkVersionVariable "$handler" true "$LINENO" || return $?
     __catchEnvironment "$handler" python -m mkdocs build || return $?
+    __checkVersionVariable "$handler" true "$LINENO" || return $?
     __catchEnvironment "$handler" muzzle popd || return $?
   fi
